+++ test/unit/group_test.rb
@@ -13,4 +13,11 @@ describe Group do
     group = create(:group)
     group.membership_for(group.user).must_be :present?
   end
+
+  it "check membership" do
+    group = create(:group)
+    user = create(:user)
+    group.add_member(user)
+    group.member?(user).must_equal true
+  end
 end