+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Masqueunacasa
 
 ## 7/05/2012
+- Bug fix: check membership (group)
 - Limpiando link_box
 - Bug: la cabecera no era responsive
 - Eliminar archivo de la barra lateral de los grupos